Adding Features
I finished the color changer tag, a couple easter eggs and some new commands. Anyone who feels like making their own branch of this for whatever reason, and you want to merge it, make it as a separate core file. I can put it as the BETA MODE. I can use that to showcase other renditions of my code.