Fixed help menu
FINALLY I got around to putting a pause after the help text. Don't listen to the pre-built macros, they don't exist yet.

On an update for the program macro, I plan on setting out a bunch of built in directories for the better known programs, but you'll have to enter in the directories yourself if you want to use something I don't know about. Feel free to add directories to the soon to be very long list of other programs, and send them to me.